tasks/requirements.txt
# Copyright (c) 2020 ASTRID
# author: Alex Carrega <alessandro.carrega@cnit.it>

docker
emoji
gitpython
invoke
rich
screenutils
python-telegram-bot
